requirements.txt
mysql-connector==2.2.9
requests==2.31.0